// src/types/notification.d.ts
export type DatabaseNotification<Config extends BaseNotificationTypeConfig> = {
  id: Config['NotificationIdType'];
  userId: Config['UserIdType'];
  notificationType: NotificationType;
  title: string | null;
  bodyTemplate: string;
  contextName: string & keyof Config['ContextMap'];
  contextParameters: Parameters<
    Config['ContextMap'][DatabaseNotification<Config>['contextName']]['generate']
  >[0];
  sendAfter: Date | null;
  subjectTemplate: string | null;
  status: NotificationStatus;
  contextUsed:
    | null
    | (ReturnType<
        Config['ContextMap'][DatabaseNotification<Config>['contextName']]['generate']
      > extends Promise<infer T>
        ? T
        : ReturnType<
            Config['ContextMap'][DatabaseNotification<Config>['contextName']]['generate']
          >);
  extraParams: JsonValue;
  adapterUsed: string | null;
  sentAt: Date | null;
  readAt: Date | null;
  createdAt?: Date;
  updatedAt?: Date;
  gitCommitSha: string | null;
  attachments?: StoredAttachment[];
};
// ...
export type AnyDatabaseNotification<Config extends BaseNotificationTypeConfig> =
  | DatabaseNotification<Config>
  | DatabaseOneOffNotification<Config>;